+++ CHANGELOG.md
@@ -4,12 +4,11 @@
 - **BF**: Getting "Hai già effettuato una chiusura" when booking time off one hour before a previous time off
 - **BC**: Change ecs ami to a free one
 - **BC**: Integrate with email automation service
+- **FR**: Add proper Favicon.ico
 - **Chapter**: Hide Repositories behind Services and Controllers
-   - Remove Rest API from PersonalTrainingBundleSpecification Repository
    - Remove Rest API from Reservation Repository
    - Remove Rest API from Time Off Repository
    - Remove Rest API from TrainingBundle Repository
-   - Remove Rest API from TrainingBundleSpecification Repository
    - Remove Rest API from TrainingBundleSession Repository
   
 - **FR**: Schedule Training Assignment
@@ -33,6 +32,9 @@
    - Removed Rest API on User Repository
    - Removed Rest API from Gym Repository
    - Added Gym Controller
+   - Removed PersonalTrainingBundleSpecification Repository
+   - Removed Rest API from TrainingBundleSpecification Repository
+   - Added TrainingBundleSpecification Service w tests
    
 - **BF**: Email not valid
 - **BF**: Typo "Domendica" in gym configuration